refactor(core): rename CardsTable to CardTable
affects: @medly-components/core
