+++ packages/core/src/components/CardTable/Body/Cell/types.ts
