+++ packages/core/src/components/CardTable/CardTable.tsx
@@ -4,14 +4,14 @@ import * as Styled from './CardTable.styled';
 import Head from './Head';
 import { Props } from './types';
 
-export const CardsTable: SFC<Props> = React.memo(({ withWhiteBackground, onSort, columns, ...restProps }) => (
+export const CardTable: SFC<Props> = React.memo(({ withWhiteBackground, onSort, columns, ...restProps }) => (
     <Styled.Table>
         <Head onSort={onSort} columns={columns} withWhiteBackground={withWhiteBackground} />
         <Body {...restProps} columns={columns} withWhiteBackground={withWhiteBackground} />
     </Styled.Table>
 ));
 
-CardsTable.displayName = 'CardsTable';
-CardsTable.defaultProps = {
+CardTable.displayName = 'CardTable';
+CardTable.defaultProps = {
     withWhiteBackground: false
 };
