+++ packages/core/src/components/CardTable/Head/Head.styled.tsx
