+++ packages/core/src/components/CardTable/Head/HeadCell/types.ts
