+++ packages/core/src/components/CardTable/Head/types.ts
