+++ packages/core/src/components/CardTable/docs/GlobalStyle.styled.tsx
