+++ packages/core/src/components/CardTable/index.ts
@@ -1,0 +1,3 @@
+import { CardTable } from './CardTable';
+
+export default CardTable;
