--- packages/core/src/components/CardsTable/index.ts
@@ -1,3 +1,0 @@
-import { CardsTable } from './CardsTable';
-
-export default CardsTable;
